backend/requirements.txt

# Core Flask Dependencies
Flask==3.1.2
flask-cors==6.0.1
pymongo==4.15.1
python-dotenv==1.1.1
gunicorn==21.2.0
werkzeug==3.1.3
cerberus==1.3.7

# JWT Authentication
PyJWT==2.10.1

# Document Processing (CPU-Only versions)
PyMuPDF==1.24.9
python-docx==0.8.11
pillow==10.2.0
numpy>=1.26.0
reportlab==4.1.0  # For PDF generation

# OCR Engines (CPU-Only)
easyocr==1.7.1
pytesseract==0.3.10
opencv-python-headless==4.10.0.84

# AI Services
google-generativeai==0.3.2
openai==1.35.7

# Phase 2: Image Captioning & Advanced Vision
transformers==4.36.0
torch==2.2.0
torchvision==0.17.0
accelerate==0.25.0
sentencepiece==0.2.1

# Phase 3: Advanced Document Understanding
# (uses transformers from Phase 2)

# Phase 4: Table Extraction
camelot-py[cv]==0.11.0
tabula-py==2.9.0
pandas==2.1.4
openpyxl==3.1.2

# Google Services (Optional)
google-auth==2.23.4
google-api-python-client==2.137.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Development & Testing
pytest==8.3.2
pytest-flask==1.3.0
pytest-mock==3.14.0
requests==2.32.3

# Utilities
python-magic==0.4.27
validators==0.22.0
